NickDev: 1. exrd.f suggests a modified way to deal with the unused dummy variablesi; this change is required to deal with cases where the dummy arguments are not defined by the caller because in that case changing the values to real causes an exception. 2. Fixed some sourceList files. 3. Modified the makefile to suppress the integer division warning. 4. Uncommented lines that left toupper() undefined in bur_brn.c and fof_unix.c ... Also cast some long ints to ints for output.
git-svn-id: svn+ssh://svn.code.sf.net/p/open-fvs/code/branches/NickDev@2158 90b89b48-9230-0de4-5065-812dd12741b8

+++ fire/fofem/src/bur_brn.c
@@ -10,7 +10,7 @@
 #include <stdio.h>
 #include <math.h>
 #include <string.h>
-/* #include <ctype.h> */
+#include <ctype.h>
 
 #include  "fof_gen.h"
 
@@ -788,7 +788,7 @@ char cr[200];
    strcpy (cr_ErrMes,"");
 
    if ( ntimes <= 0 ) {
-     sprintf(cr_ErrMes, "Burnup: Number of iterations invalid, %d, must be > 0",ntimes);
+     sprintf(cr_ErrMes, "Burnup: Number of iterations invalid, %d, must be > 0",(int) ntimes);
      return false; }
 
    for (i = 0; i < number; i++) {
